+++ front/assets/html/cgu.html
@@ -1,10 +1,10 @@
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
+<!DOCTYPE html>
 <html>
   <head>
     <meta http-equiv="CONTENT-TYPE" content="text/html; charset=utf-8" />
     <title></title>
   </head>
-  <body lang="fr-FR">
+  <body>
 
     <!---------------------------------------------------->
     <!-- !! LAISSER TOUTES LES BALISES SUR UNE LIGNE !! -->
