+++ front/components/Modal.tsx
@@ -1,19 +1,49 @@
+import table, { IGNORED_TAGS } from "@native-html/table-plugin";
+import type { Asset } from "expo-asset";
 import * as React from "react";
+import { useEffect } from "react";
 import { Modal as RNModal, SafeAreaView, StyleSheet } from "react-native";
 import { ScrollView } from "react-native-gesture-handler";
+import HTML from "react-native-render-html";
+import { WebView } from "react-native-webview";
 
-import { Paddings } from "../constants";
+import { FontWeight, Paddings } from "../constants";
+import { AssestUtils } from "../utils";
 import { Button } from ".";
 import { View } from "./Themed";
 
 interface Props {
-  content: JSX.Element;
+  content?: React.ReactNode;
+  htmlFile?: string;
   setIsVisible: (showMenu: boolean) => void;
 }
 
-const Modal: React.FC<Props> = ({ content, setIsVisible }) => {
+const Modal: React.FC<Props> = ({ content, setIsVisible, htmlFile }) => {
   const hideModal = () => {
     setIsVisible(false);
+  };
+  const [html, setHtml] = React.useState<Asset | null>(null);
+
+  useEffect(() => {
+    if (htmlFile) {
+      void AssestUtils.getAssetsAsync(htmlFile).then((asset: Asset) => {
+        setHtml(asset);
+      });
+    }
+  }, []);
+
+  const buildContent = (data: Asset) => {
+    const htmlProps = {
+      // eslint-disable-next-line @typescript-eslint/naming-convention
+      WebView,
+      ignoredTags: IGNORED_TAGS,
+      renderers: {
+        table,
+      },
+    };
+    return (
+      <HTML source={data} tagsStyles={{ b: styles.bold }} {...htmlProps} />
+    );
   };
 
   return (
@@ -28,7 +58,8 @@ const Modal: React.FC<Props> = ({ content, setIsVisible }) => {
           <View style={styles.closeButton}>
             <Button title="X" rounded={true} action={hideModal} />
           </View>
-          <View>{content}</View>
+          {content && <View>{content}</View>}
+          {html && <View>{buildContent(html)}</View>}
         </ScrollView>
       </SafeAreaView>
     </RNModal>
@@ -36,6 +67,9 @@ const Modal: React.FC<Props> = ({ content, setIsVisible }) => {
 };
 
 const styles = StyleSheet.create({
+  bold: {
+    fontWeight: FontWeight.bold,
+  },
   closeButton: {
     alignSelf: "flex-end",
   },
