+++ front/utils/asset.util.tsx
@@ -4,7 +4,5 @@ import * as AssetUtils from "expo-asset-utils";
 export const getAssetsAsync = async (asset: string): Promise<Asset> => {
   // eslint-disable-next-line @typescript-eslint/no-unsafe-argument
   const expoAsset = await Asset.fromModule(asset).downloadAsync();
-  return AssetUtils.resolveAsync(expoAsset).then((newAsset: Asset) => {
-    return newAsset;
-  });
+  return AssetUtils.resolveAsync(expoAsset).then((newAsset: Asset) => newAsset);
 };
